requirements.txt
quart
quart-cors
aiohttp
python-dotenv
requests